Keep localization in the fenix search engine provider
